requirements.txt
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.8.2
tqdm==4.66.4
streamlit==1.39.0
pytest==8.2.0